exercises/requirements.txt
django==1.3.1
django-mptt
